Move tokens API endpoint to controller.
