+++ includes/classes/Feature/WooCommerce/OrdersAutosuggest.php
@@ -10,6 +10,7 @@ namespace ElasticPress\Feature\WooCommerce;
 
 use ElasticPress\Elasticsearch;
 use ElasticPress\Indexables;
+use ElasticPress\REST;
 use ElasticPress\Utils;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -99,44 +100,13 @@ class OrdersAutosuggest {
 	}
 
 	/**
-	 * Get the endpoint for temporary tokens.
-	 *
-	 * @return string Temporary token endpoint.
-	 */
-	public function get_token_endpoint() {
-		/**
-		 * Filters the temporary token API endpoint.
-		 *
-		 * @since 4.5.0
-		 * @hook ep_token_endpoint
-		 * @param {string} $endpoint Endpoint path.
-		 * @returns {string} Token API endpoint.
-		 */
-		return apply_filters( 'ep_token_endpoint', 'api/v1/token' );
-	}
-
-	/**
 	 * Registers the API endpoint to get a token.
 	 *
 	 * @return void
 	 */
 	public function rest_api_init() {
-		register_rest_route(
-			'elasticpress/v1',
-			'token',
-			[
-				[
-					'callback'            => [ $this, 'get_token' ],
-					'permission_callback' => [ $this, 'check_token_permission' ],
-					'methods'             => 'GET',
-				],
-				[
-					'callback'            => [ $this, 'refresh_token' ],
-					'permission_callback' => [ $this, 'check_token_permission' ],
-					'methods'             => 'POST',
-				],
-			]
-		);
+		$controller = new REST\Token();
+		$controller->register_routes();
 	}
 
 	/**
@@ -389,66 +359,6 @@ class OrdersAutosuggest {
 		);
 
 		return $args;
-	}
-
-	/**
-	 * Get a temporary token.
-	 *
-	 * @return string|false Authorization header, or false on failure.
-	 */
-	public function get_token() {
-		$user_id = get_current_user_id();
-
-		$credentials = get_user_meta( $user_id, 'ep_token', true );
-
-		if ( $credentials ) {
-			return $credentials;
-		}
-
-		return $this->refresh_token();
-	}
-
-	/**
-	 * Refresh the temporary token.
-	 *
-	 * @return string|false Authorization header, or false on failure.
-	 */
-	public function refresh_token() {
-		$user_id = get_current_user_id();
-
-		$endpoint = $this->get_token_endpoint();
-		$response = Elasticsearch::factory()->remote_request( $endpoint, [ 'method' => 'POST' ] );
-
-		if ( is_wp_error( $response ) || 200 !== wp_remote_retrieve_response_code( $response ) ) {
-			return false;
-		}
-
-		$response = wp_remote_retrieve_body( $response );
-		$response = json_decode( $response );
-
-		$credentials = base64_encode( "$response->username:$response->clear_password" ); // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_encode
-
-		update_user_meta( $user_id, 'ep_token', $credentials );
-
-		return $credentials;
-	}
-
-	/**
-	 * Checks if the token API can be used.
-	 *
-	 * @return boolean Whether the token API can be used.
-	 */
-	public function check_token_permission() {
-		/**
-		 * Filters the capability required to use the token API.
-		 *
-		 * @since 4.5.0
-		 * @hook ep_token_capability
-		 * @param {string} $capability Required capability.
-		 */
-		$capability = apply_filters( 'ep_token_capability', 'edit_others_shop_orders' );
-
-		return current_user_can( $capability );
 	}
 
 	/**
