+++ tests/php/features/WooCommerce/TestWooCommerceOrders.php
@@ -307,7 +307,6 @@ class TestWooCommerceOrders extends TestWooCommerce {
 	public function ordersAutosuggestMethodsDataProvider() : array {
 		return [
 			[ 'after_update_feature', [ 'test', [], [] ] ],
-			[ 'check_token_permission', [] ],
 			[ 'enqueue_admin_assets', [ '' ] ],
 			[ 'epio_delete_search_template', [] ],
 			[ 'epio_get_search_template', [] ],
@@ -317,8 +316,6 @@ class TestWooCommerceOrders extends TestWooCommerce {
 			[ 'get_search_endpoint', [] ],
 			[ 'get_search_template', [] ],
 			[ 'get_template_endpoint', [] ],
-			[ 'get_token', [] ],
-			[ 'get_token_endpoint', [] ],
 			[ 'intercept_search_request', [ (object) [] ] ],
 			[ 'is_integrated_request', [ true, [] ] ],
 			[ 'post_statuses', [ [] ] ],
@@ -326,7 +323,6 @@ class TestWooCommerceOrders extends TestWooCommerce {
 			[ 'mapping', [ [] ] ],
 			[ 'maybe_query_password_protected_posts', [ [] ] ],
 			[ 'maybe_set_posts_where', [ '', new \WP_Query( [] ) ] ],
-			[ 'refresh_token', [] ],
 			[ 'rest_api_init', [] ],
 			[ 'set_search_fields', [] ],
 		];
